requirements.txt
kafka-python
s3fs
tweepy
py_dotenv
requests
textblob
nltk